refactor(room): split room page hook into granular hooks
