+++ src/lib/components/room/header/types.ts
@@ -2,5 +2,4 @@ import type { RoomInstance } from "lib/types/RawDB";
 
 export type RoomHeaderProps = {
   roomData?: RoomInstance;
-  isOwner: boolean;
 };
