+++ src/lib/utils/roomUtils.ts
